Update TesSense.py
Bug fixes and added MinRate a variable that allows for charging below the 5 amps the charger defaults to. This allows the charging to continue when the free solar gets very limited for a short time keeping the charger from stopping which is likely harder on the charging hardware than simply lowering the amps output.